First overhaul of the project structure
