+++ .gitignore
@@ -1,8 +1,3 @@
-# File created using '.gitignore Generator' for Visual Studio Code: https://bit.ly/vscode-gig
-
-# Created by https://www.gitignore.io/api/macos,visualstudiocode,go
-# Edit at https://www.gitignore.io/?templates=macos,visualstudiocode,go
-
 ### Go ###
 # Binaries for programs and plugins
 *.exe
@@ -59,8 +54,3 @@ Temporary Items
 ### VisualStudioCode Patch ###
 # Ignore all local history of files
 .history
-
-# End of https://www.gitignore.io/api/macos,visualstudiocode,go
-
-# Custom rules (everything added below won't be overriden by 'Generate .gitignore File' if you use 'Update' option)
-
