+++ internal/pkg/aws-s3/get-objects.go
@@ -1,0 +1,27 @@
+package s3
+
+import (
+	"io"
+	"net/http"
+	"os"
+)
+
+func DownloadFromS3Bucket(filepath, fileUrl string) error {
+	// Get the data
+	resp, err := http.Get(fileUrl)
+	if err != nil {
+		return err
+	}
+	defer resp.Body.Close()
+
+	// Create the file
+	out, err := os.Create(filepath)
+	if err != nil {
+		return err
+	}
+	defer out.Close()
+
+	// Write the body to file
+	_, err = io.Copy(out, resp.Body)
+	return err
+}
