--- run
@@ -1,7 +1,0 @@
-DAYNUMBER=$1
-
-rm -rf tools work
-./spawn.pl ../d$DAYNUMBER config_d$DAYNUMBER.pl
-./tools/build.sh
-./tools/verify.sh
-./tools/check_all.sh
