--- tools/check_all.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-
